added in basic weather wrapper/looked into details for customization
