+++ Dreams VR/Assets/NickWeatherManager.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 0c3b1d584724d5943b5eef649074a961
+timeCreated: 1526979956
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
