+++ Dreams VR/Assets/Scripts/NickScripts/VRFootsteps.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 693d7ce03df98d942a8bbc150f7726c4
+timeCreated: 1526434181
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
